% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leven.R
\name{bestMismatch}
\alias{bestMismatch}
\title{Find fewest mismatches}
\usage{
bestMismatch(pattern, subject, pos = findAll, weights = rep(1,
  nchar(pattern)), findAll = FALSE, dummyChar = "|")
}
\arguments{
\item{pattern}{String pattern to search for in subject}

\item{subject}{Subject string to search in}

\item{pos}{If TRUE return position in addition to distance}

\item{weights}{Integer weights for each position of pattern}

\item{findAll}{If TRUE return all best matches}

\item{dummyChar}{A character that does not appear in either string (only necessary if findAll==TRUE [and could probably program better])}
}
\value{
Single value giving the least mismatch if pos=FALSE, c(minMismatches,position) if pos=TRUE, if findAll==TRUE matrix with columns minMismatches, position and rows for each occurrence
}
\description{
Find fewest mismatches (and position if desired) between pattern and string without allowing gaps. Could be programmed better but quick enough for now.
}
\author{
Scott Sherrill-Mix \email{R@sherrillmix.com}
}

