% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leven.R
\name{combineAligns}
\alias{combineAligns}
\title{Combine gapped alignments}
\usage{
combineAligns(refs, aligns, starts = NULL)
}
\arguments{
\item{refs}{The gapped reference sequences (one for each query)}

\item{aligns}{The gapped query sequences}

\item{starts}{If refs start at different locations add .'s so all start at 1}
}
\value{
A vector of the aligned strings with gaps added
}
\description{
Combine several pairs of reference, query alignments
}
\author{
Scott Sherrill-Mix \email{R@sherrillmix.com}
}

