% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leven.R
\name{complimentDna}
\alias{complimentDna}
\title{Compliment DNA}
\usage{
complimentDna(dnas)
}
\arguments{
\item{dnas}{Vector of DNA sequences}
}
\value{
Vector of complimented DNA strings
}
\description{
Compliment a string of DNA e.g. A<->T,G<->C. Doesn't attempt to deal with ambigous bases.
}

