% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leven.R
\name{.hamming}
\alias{.hamming}
\title{Wrapper for .C Hamming function}
\usage{
.hamming(string1, string2, cutoff = 99999)
}
\arguments{
\item{string1}{A single string or vector of strings}

\item{string2}{Another single string or vector of strings}

\item{cutoff}{an integer giving a distance to stop checking beyond and instead return the cutoff value}
}
\value{
the hamming distance between two strings
}
\description{
Align and calculate distance betweeen two strings based on Hamming distance
}
\examples{
levenR:::.hamming('aazaazaaz','aayaayaay')
levenR:::.hamming('aazaazaaz','aayaayaay',cutoff=2)
levenR:::.hamming('ss','ss')
}
\references{
\url{http://en.wikipedia.org/wiki/Hamming_distance}
}
\author{
Scott Sherrill-Mix \email{R@sherrillmix.com}
}
