% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leven.R
\name{hamming}
\alias{hamming}
\title{Find hamming distance between many strings}
\usage{
hamming(strings1, strings2 = NULL, oneToOne = FALSE, cutoff = 99999, vocal = 0)
}
\arguments{
\item{strings1}{vector of strings to compare to strings2 (or as strings1 x strings1 if strings2==NULL)}

\item{strings2}{vector of strings to compare to strings1 or NULL (either strings1 or strings2 should be 1 or should have equal lengths)}

\item{oneToOne}{compare strings1[1] to strings2[1], strings1[2] to strings2[2],... (both vectors must be same length)}

\item{cutoff}{an integer giving a distance to stop checking beyond and instead return the cutoff value}

\item{vocal}{Integer to display a status message every vocal iterations}
}
\value{
Hamming distance matrix
}
\description{
Find hamming distance between many strings
}
\references{
\url{http://en.wikipedia.org/wiki/Hamming_distance}
}
\author{
Scott Sherrill-Mix \email{R@sherrillmix.com}
}
