% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leven.R
\name{leven}
\alias{leven}
\title{Find Levenshtein distance between many strings}
\usage{
leven(
  strings1,
  strings2 = NULL,
  oneToOne = FALSE,
  homoLimit = 0,
  vocal = 0,
  prepend = NULL,
  append = NULL,
  substring1 = FALSE,
  substring2 = FALSE,
  nThreads = 1
)
}
\arguments{
\item{strings1}{vector of strings to compare to strings2 (or as strings1 x strings1 if strings2==NULL)}

\item{strings2}{vector of strings to compare to strings1 or NULL (either strings1 or strings2 should be 1 or should have equal lengths)}

\item{oneToOne}{compare strings1[1] to strings2[1], strings1[2] to strings2[2],... (both vectors must be same length)}

\item{homoLimit}{deletions or insertions in homopolymers > homoLimit cost 0}

\item{vocal}{Integer to display a status message every vocal iterations}

\item{prepend}{1 or 2 for ends-free on front of string 1 or 2}

\item{append}{1 or 2 for ends-free on back of string 1 or 2}

\item{substring1}{ends-free for strings1}

\item{substring2}{ends-free for strings2}

\item{nThreads}{Run using threads?}
}
\value{
Levenshtein distance matrix
}
\description{
Find Levenshtein distance between many strings
}
\references{
\url{http://en.wikipedia.org/wiki/Levenshtein_distance}
}
\author{
Scott Sherrill-Mix \email{R@sherrillmix.com}
}
