% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leven.R
\name{levenAlign}
\alias{levenAlign}
\title{Align many strings to a reference}
\usage{
levenAlign(strings, ref, homoLimit = 0, prepend = NULL, append = NULL,
  substring1 = FALSE, substring2 = FALSE, trimOuterGaps = FALSE,
  revComp = FALSE, ...)
}
\arguments{
\item{strings}{The query strings to be aligned}

\item{ref}{The reference to align to}

\item{homoLimit}{deletions or insertions in homopolymers > homoLimit cost 0}

\item{prepend}{1 or 2 for ends-free on front of string 1 or 2}

\item{append}{1 or 2 for ends-free on back of string 1 or 2}

\item{substring1}{ends-free for query strings}

\item{substring2}{ends-free for reference}

\item{trimOuterGaps}{Trim leading and trailing gaps}

\item{revComp}{Find and return the best alignment of identity or reverse compliment of queries}

\item{...}{Additional arguments for levenAll}
}
\value{
A list with first entry for the gapped reference and second entry all the aligned strings
}
\description{
Align strings to a reference string based on Levenshtein distance with ends-free or homopolymer alignments possible
}
\author{
Scott Sherrill-Mix \email{R@sherrillmix.com}
}
\references{
\url{http://en.wikipedia.org/wiki/Levenshtein_distance}
}

