% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leven.R
\docType{package}
\name{levenR-package}
\alias{levenR-package}
\alias{levenstein}
\title{Simple Levenshtein alignments}
\description{
Simple Levenshtein alignments
}
\details{
\tabular{ll}{
Package: \tab levenR\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2011-11-16\cr
License: \tab CC BY-NC 3.0\cr
LazyLoad: \tab yes\cr
}
}
\examples{
library(levenR)
leven('ACAC','ACACA')
levenAlign(c('ACCCTG','AACTG'),'ACTG')
}
\references{
\url{http://en.wikipedia.org/wiki/Levenshtein_distance}
}
\seealso{
\code{Biostrings}
}
\author{
Scott Sherrill-Mix \email{R@sherrillmix.com}
}
\keyword{alignment,}
\keyword{distance,}
\keyword{levenshtein}
\keyword{string,}
