% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leven.R
\name{multiMismatch}
\alias{multiMismatch}
\title{Find fewest mismatches for multiple patterns}
\usage{
multiMismatch(patterns, subject, drop = TRUE, ...)
}
\arguments{
\item{patterns}{A vector of string patterns to search for in subject}

\item{subject}{A single subject string to search in}

\item{drop}{If TRUE return only return the first best match}

\item{...}{Additional arguments for bestMismatch}
}
\value{
Three column matrix with columns minMismatch, position, pattern number
}
\description{
Find fewest mismatches (and position if desired) between pattern and string without allowing gaps. Could be programmed better but quick enough for now
}
\author{
Scott Sherrill-Mix \email{R@sherrillmix.com}
}
