% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leven.R
\name{revComp}
\alias{revComp}
\title{Reverse compliment dna}
\usage{
revComp(dnas)
}
\arguments{
\item{dnas}{Vector of sequences}
}
\value{
Vector of reverse complimented DNA strings
}
\description{
Reverse compliment a string of DNA
}
