% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/leven.R
\name{reverseString}
\alias{reverseString}
\title{Reverse a string}
\usage{
reverseString(strings)
}
\arguments{
\item{strings}{Vector of strings to be revered}
}
\value{
Vector of reversed strings
}
\description{
Reverse a string e.g. ABCDE becomes EDCBA
}

